feat: add shortcut grid widget with drag-and-drop functionality
- Implemented ShortcutIcon component to display shortcut icons with brand and image support.
- Created index file for shortcut icons export.
- Added CSS styles for shortcut tiles, including hover effects and drag-and-drop visuals.
- Developed useGridDragSession hook to manage drag-and-drop interactions within the shortcut grid.
- Introduced useNativeDragOverlay for handling drag overlay positioning.
- Added widget-frame styles for visual representation of widgets during editing.
- Created a main widgets CSS file to import various widget styles.
