+++ docs/architecture-review.md
@@ -1,6 +1,6 @@
 # Architecture Review
 
-Latest pass: 2026-05-06.
+Latest pass: 2026-05-12.
 
 ## Fixed In This Pass
 
@@ -12,12 +12,21 @@ Latest pass: 2026-05-06.
 - Grid metric measurement reacts to column spacing changes.
 - Cross-page Top-Level Tile drag now uses page-edge hover navigation close to the reference extension.
 - Docs now reflect React 19, global CSS, active Zustand persistence, FolderPanel child drag, release automation, and current architecture docs.
+- UI files now follow product-concept locality: Canvas hosts Widgets, Widget Modules own their rendering/settings menu sections, Popup owns Toolbar Popup composition, and shared Shortcut UI owns forms/icons/tile styling.
+- `src/domain/drafts.ts` owns edit draft contracts so domain operations no longer import UI model files.
+- `shortcutPageModel.ts` owns pure Shortcut Page derivation and insertion-position math for Grid Layout Capacity.
+- `src/ui/styles.css` is a global stylesheet index; Module-owned selectors live beside their UI Modules.
+- Removed stale New Tab Surface Shortcut creation plumbing from the Shortcut Grid Widget and FolderPanel; the Toolbar Popup remains the Shortcut creation flow.
+- Renumbered the native drag ADR to avoid duplicate ADR `0005` identifiers.
+- Removed inactive dnd-kit dependencies so native drag is the only installed drag Adapter.
+- Removed duplicate Search query state from `useNewTabController`; the New Tab Surface owns transient query input state.
+- Added an explicit Canvas Widget controller Interface so `CanvasWidgetHost` no longer depends on the full controller Implementation type.
 
 ## Remaining Deepening Opportunities
 
 1. **Drag Session Module**
 
-Files: `src/ui/ShortcutGrid.tsx`, `src/ui/modals/FolderPanel.tsx`, `src/ui/drag/*`.
+Files: `src/ui/widgets/shortcut-grid/ShortcutGrid.tsx`, `src/ui/widgets/shortcut-grid/FolderPanel.tsx`, `src/ui/drag/*`.
 
 Problem: drag session timers, overlays, geometry, and action dispatch still spread across UI surfaces.
 
@@ -47,7 +56,7 @@ Benefit: clearer persistence interface and docs; fewer false assumptions about s
 
 4. **Overlay Focus Module**
 
-Files: `src/ui/SettingsDrawer.tsx`, `src/ui/modals/*`.
+Files: `src/ui/settings/SettingsDrawer.tsx`, `src/ui/modals/*`.
 
 Problem: modal/drawer overlays set `aria-modal` but do not trap focus or restore focus.
