+++ package.json
@@ -11,9 +11,6 @@
     "test:smoke": "playwright test"
   },
   "dependencies": {
-    "@dnd-kit/core": "^6.3.1",
-    "@dnd-kit/sortable": "^10.0.0",
-    "@dnd-kit/utilities": "^3.2.2",
     "@vitejs/plugin-react": "^5.0.4",
     "immer": "^11.1.4",
     "lucide-react": "^1.14.0",
