+++ src/popup.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { PopupApp } from "./ui/PopupApp";
+import { PopupApp } from "./ui/popup";
 import "./ui/styles.css";
 
 createRoot(document.getElementById("root")!).render(
