+++ src/ui/app/App.tsx
@@ -1,44 +1,23 @@
-import { CSSProperties, MouseEvent, useEffect, useState } from "react";
+import { CSSProperties, useState } from "react";
 import { defaultTabState } from "../../domain/tabState";
 import type { DragSource } from "../drag/dragModel";
-import { CanvasSurface, WidgetFrame } from "../canvas/CanvasSurface";
-import { SettingsDrawer } from "../SettingsDrawer";
-import { WidgetContextMenu, type ContextMenuState } from "../widgets/WidgetContextMenu";
-import { SearchWidget } from "../widgets/search/SearchWidget";
-import { ShortcutGridWidget } from "../widgets/shortcut-grid/ShortcutGridWidget";
-import { FolderModal } from "../modals/FolderModal";
-import { FolderPanel } from "../modals/FolderPanel";
-import { ShortcutModal } from "../modals/ShortcutModal";
+import { CanvasWidgetHost } from "../canvas";
+import { SettingsDrawer } from "../settings";
+import { FolderModal, ShortcutModal } from "../modals";
+import { FolderPanel } from "../widgets/shortcut-grid";
 import { useNewTabController } from "./useNewTabController";
-import { useCanvasMetrics } from "../canvas/useCanvasMetrics";
+import { useCanvasMetrics } from "../canvas";
 
 export function App() {
   const controller = useNewTabController();
   const [query, setQuery] = useState("");
   const [isCanvasEditMode, setIsCanvasEditMode] = useState(false);
-  const [contextMenu, setContextMenu] = useState<ContextMenuState | null>(null);
   const [outgoingDragSource, setOutgoingDragSource] = useState<DragSource | null>(null);
 
   const tabState = controller.tabState ?? defaultTabState;
   const canvasMetrics = useCanvasMetrics(tabState.canvas.targetCellSize);
-  const shortcutGridWidget = tabState.canvas.widgets.shortcutGrid;
   const searchWidget = tabState.canvas.widgets.search;
   const searchSettings = searchWidget.settings;
-
-  useEffect(() => {
-    if (!contextMenu) {
-      return;
-    }
-
-    function closeContextMenu(event: KeyboardEvent) {
-      if (event.key === "Escape") {
-        setContextMenu(null);
-      }
-    }
-
-    window.addEventListener("keydown", closeContextMenu);
-    return () => window.removeEventListener("keydown", closeContextMenu);
-  }, [contextMenu]);
 
   const searchBoxHeight = Math.max(44, searchWidget.placement.height * canvasMetrics.cellHeight);
   const searchBoxRoundness = Math.min(100, Math.max(0, searchSettings.radius));
@@ -56,25 +35,6 @@ export function App() {
     "--wallpaper-dim": `${tabState.wallpaper.dim / 100}`,
     "--wallpaper-blur": `${tabState.wallpaper.blur}px`
   } as CSSProperties;
-
-  function openCanvasContextMenu(event: MouseEvent<HTMLElement>) {
-    if (!isCanvasEditMode || event.target !== event.currentTarget) {
-      return;
-    }
-
-    event.preventDefault();
-    setContextMenu({ type: "canvas", x: event.clientX, y: event.clientY });
-  }
-
-  function openWidgetContextMenu(widgetId: "search" | "shortcutGrid", event: MouseEvent<HTMLElement>) {
-    if (!isCanvasEditMode) {
-      return;
-    }
-
-    event.preventDefault();
-    event.stopPropagation();
-    setContextMenu({ type: "widget", widgetId, x: event.clientX, y: event.clientY });
-  }
 
   if (!controller.tabState) {
     return <main className="new-tab loading">Loading</main>;
@@ -111,71 +71,17 @@ export function App() {
           </button>
         </div>
 
-        <CanvasSurface editMode={isCanvasEditMode} metrics={canvasMetrics} onCanvasContextMenu={openCanvasContextMenu}>
-          <WidgetFrame
-            editMode={isCanvasEditMode}
-            enabled={searchWidget.enabled}
-            label="Search Widget"
-            metrics={canvasMetrics}
-            onMove={(placement) => controller.updateWidgetPlacement("search", placement, canvasMetrics)}
-            onResize={(placement) => controller.updateWidgetPlacement("search", placement, canvasMetrics)}
-            onWidgetContextMenu={openWidgetContextMenu}
-            placement={searchWidget.placement}
-            widgetId="search"
-          >
-            <SearchWidget
-              activeProvider={controller.activeSearchProvider}
-              changeSearchProvider={controller.changeSearchProvider}
-              query={query}
-              setQuery={setQuery}
-              settings={searchSettings}
-            />
-          </WidgetFrame>
-
-          <WidgetFrame
-            editMode={isCanvasEditMode}
-            enabled={shortcutGridWidget.enabled}
-            label="Shortcut Grid Widget"
-            metrics={canvasMetrics}
-            onMove={(placement) => controller.updateWidgetPlacement("shortcutGrid", placement, canvasMetrics)}
-            onResize={(placement) => controller.updateWidgetPlacement("shortcutGrid", placement, canvasMetrics)}
-            onWidgetContextMenu={openWidgetContextMenu}
-            placement={shortcutGridWidget.placement}
-            widgetId="shortcutGrid"
-          >
-            <ShortcutGridWidget
-              activeShortcutPage={controller.activeShortcutPage}
-              canvasMetrics={canvasMetrics}
-              dispatchDropAction={controller.dispatchDropAction}
-              gridRef={controller.gridRef}
-              hasOverlayOpen={controller.hasOverlayOpen}
-              isCanvasEditMode={isCanvasEditMode}
-              onClearOutgoingDrag={() => setOutgoingDragSource(null)}
-              onEditFolder={controller.openEditFolderDialog}
-              onEditShortcut={controller.openEditShortcutDialog}
-              onOpenNewShortcutDialog={controller.openNewShortcutDialog}
-              onSetActiveFolderId={controller.setActiveFolderId}
-              outgoingDragSource={outgoingDragSource}
-              setActiveShortcutPage={controller.setActiveShortcutPage}
-              settings={shortcutGridWidget.settings}
-              tabState={tabState}
-              topLevelTiles={controller.topLevelTiles}
-              widgetPlacement={shortcutGridWidget.placement}
-            />
-          </WidgetFrame>
-        </CanvasSurface>
-      </section>
-
-      {contextMenu ? (
-        <WidgetContextMenu
-          changeSearchWidgetSetting={controller.changeSearchWidgetSetting}
-          changeShortcutGridWidgetSetting={controller.changeShortcutGridWidgetSetting}
-          close={() => setContextMenu(null)}
-          menu={contextMenu}
-          setWidgetEnabled={(widgetId, enabled) => controller.setWidgetEnabled(widgetId, enabled, canvasMetrics)}
+        <CanvasWidgetHost
+          canvasMetrics={canvasMetrics}
+          controller={controller}
+          isCanvasEditMode={isCanvasEditMode}
+          outgoingDragSource={outgoingDragSource}
+          query={query}
+          setOutgoingDragSource={setOutgoingDragSource}
+          setQuery={setQuery}
           tabState={tabState}
         />
-      ) : null}
+      </section>
 
       {controller.isSettingsDrawerOpen ? (
         <SettingsDrawer
@@ -199,7 +105,6 @@ export function App() {
           onClose={() => controller.setActiveFolderId(null)}
           onEditFolder={controller.openEditFolderDialog}
           onEditShortcut={(shortcut) => controller.openEditShortcutDialog(shortcut, controller.activeFolder?.id ?? null)}
-          onOpenNewShortcutDialog={controller.openNewShortcutDialog}
           onStartOutgoingDrag={(source) => {
             setOutgoingDragSource(source);
           }}
