+++ src/ui/app/app.css
@@ -1,0 +1,116 @@
+.new-tab {
+  --icon-size: 86px;
+  --grid-column-gap: 34px;
+  --grid-row-gap: 34px;
+  --grid-columns: 6;
+  --grid-rows: 2;
+  --quick-link-label-font-size: 16px;
+  --quick-link-tile-gap: 12px;
+  --wallpaper-dim: 0.4;
+  --wallpaper-blur: 0px;
+  min-height: 100vh;
+  overflow: hidden;
+  position: relative;
+  background: #283333;
+}
+
+.new-tab.loading {
+  display: grid;
+  place-items: center;
+  color: rgba(255, 255, 255, 0.76);
+}
+
+.wallpaper {
+  position: absolute;
+  inset: 0;
+  overflow: hidden;
+  background: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1600' height='1000' viewBox='0 0 1600 1000'%3E%3Crect width='1600' height='1000' fill='%235c665f'/%3E%3Cpath d='M0 754c180-92 320-102 482-32 160 68 310 52 468-10 254-100 431-72 650 50v238H0z' fill='%23222f2e'/%3E%3Cpath d='M0 652c154-82 302-120 455-80 170 45 262 120 462 60 235-70 449-138 683 8v360H0z' fill='%2332463e' opacity='.72'/%3E%3Cg fill='%23172424' opacity='.72'%3E%3Cpath d='M120 845l44-172 44 172h-25v85h-39v-85z'/%3E%3Cpath d='M310 820l54-220 55 220h-31v122h-49V820z'/%3E%3Cpath d='M520 838l44-166 44 166h-24v100h-40V838z'/%3E%3Cpath d='M1225 828l48-190 49 190h-27v108h-44V828z'/%3E%3Cpath d='M1420 852l38-150 38 150h-21v80h-34v-80z'/%3E%3C/g%3E%3C/svg%3E");
+  background-size: cover;
+  background-position: center;
+}
+
+.wallpaper::after {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background: rgba(0, 0, 0, var(--wallpaper-dim));
+}
+
+.wallpaper-media {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  filter: blur(var(--wallpaper-blur));
+  transform: scale(1.04);
+}
+
+.workspace {
+  position: relative;
+  z-index: 1;
+  display: grid;
+  height: 100vh;
+  min-height: 0;
+  grid-template-rows: auto auto minmax(0, 1fr) 56px;
+  justify-items: center;
+  padding: 28px clamp(20px, 4vw, 72px) 58px;
+  overflow: hidden;
+}
+
+.toolbar {
+  display: flex;
+  width: 100%;
+  grid-row: 1;
+  gap: 10px;
+  justify-content: flex-end;
+  position: relative;
+  z-index: 30;
+}
+
+.edit-mode-button {
+  height: 50px;
+  border: 1px solid rgba(255, 255, 255, 0.18);
+  border-radius: 999px;
+  background: rgba(22, 27, 27, 0.48);
+  color: #ffffff;
+  cursor: pointer;
+  font-weight: 800;
+  letter-spacing: 0.02em;
+  padding: 0 18px;
+  box-shadow: 0 10px 24px rgba(0, 0, 0, 0.18);
+}
+
+.edit-mode-button:hover,
+.edit-mode-button.active {
+  background: rgba(96, 165, 250, 0.8);
+}
+
+.settings-button {
+  display: grid;
+  width: 50px;
+  height: 50px;
+  place-items: center;
+  border: 0;
+  border-radius: 50%;
+  background: rgba(22, 27, 27, 0.48);
+  color: #ffffff;
+  cursor: pointer;
+  box-shadow: 0 10px 24px rgba(0, 0, 0, 0.18);
+  font-size: 23px;
+}
+
+.settings-button:hover {
+  background: rgba(22, 27, 27, 0.62);
+}
+
+.gear-icon {
+  display: block;
+  line-height: 1;
+}
+
+@media (max-width: 760px) {
+  .workspace {
+    padding: 20px 18px 42px;
+  }
+}
