+++ src/ui/base.css
@@ -1,0 +1,33 @@
+:root {
+  color: #f8fafc;
+  font-family:
+    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+html,
+body,
+#root {
+  height: 100%;
+  margin: 0;
+  overflow: hidden;
+}
+
+body.popup-body,
+body.popup-body #root {
+  width: 420px;
+  height: auto;
+  min-height: 0;
+  overflow: auto;
+  background: #eef2f7;
+}
+
+button,
+input {
+  font: inherit;
+}
