+++ src/ui/canvas/canvas.css
@@ -1,0 +1,6 @@
+.canvas-surface {
+  position: absolute;
+  z-index: 2;
+  inset: 0;
+  overflow: hidden;
+}
