+++ src/ui/canvas/index.ts
@@ -1,0 +1,2 @@
+export { CanvasWidgetHost } from "./CanvasWidgetHost";
+export { useCanvasMetrics } from "./useCanvasMetrics";
