+++ src/ui/modals/ShortcutModal.tsx
@@ -1,7 +1,7 @@
 import type { FormEvent } from "react";
 import { type BrandIcon } from "../../domain/brandIcons";
-import { ShortcutForm } from "../ShortcutForm";
-import { type ShortcutDraft } from "../model/drafts";
+import { type ShortcutDraft } from "../../domain/drafts";
+import { ShortcutForm } from "../shortcut-editor";
 
 type ShortcutModalProps = {
   draft: ShortcutDraft;
