+++ src/ui/modals/index.ts
@@ -1,0 +1,2 @@
+export { FolderModal } from "./FolderModal";
+export { ShortcutModal } from "./ShortcutModal";
