+++ src/ui/popup/PopupApp.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useMemo, useState, type FormEvent } from "react";
-import { findBrandIconRecommendations, type BrandIcon } from "../domain/brandIcons";
-import { applyRecommendedIcon, createShortcutFromDraft, upsertShortcut } from "../domain/tabOperations";
-import { readFileAsDataUrl } from "../infrastructure/fileData";
-import { useTabStore } from "../stores/useTabStore";
-import { ShortcutForm } from "./ShortcutForm";
-import { emptyShortcutDraft, type ShortcutDraft } from "./model/drafts";
+import { findBrandIconRecommendations, type BrandIcon } from "../../domain/brandIcons";
+import { emptyShortcutDraft, type ShortcutDraft } from "../../domain/drafts";
+import { applyRecommendedIcon, createShortcutFromDraft, upsertShortcut } from "../../domain/tabOperations";
+import { readFileAsDataUrl } from "../../infrastructure/fileData";
+import { useTabStore } from "../../stores/useTabStore";
+import { ShortcutForm } from "../shortcut-editor";
 
 type ActiveTab = {
   title?: string;
