+++ src/ui/popup/index.ts
@@ -1,0 +1,1 @@
+export { PopupApp } from "./PopupApp";
