+++ src/ui/settings/settings.css
@@ -1,0 +1,220 @@
+.settings-drawer-backdrop {
+  position: fixed;
+  z-index: 40;
+  inset: 0;
+  display: flex;
+  justify-content: flex-end;
+  background: rgba(2, 6, 23, 0.36);
+  backdrop-filter: blur(8px);
+}
+
+.settings-drawer {
+  display: grid;
+  width: min(400px, calc(100vw - 34px));
+  height: 100vh;
+  grid-template-rows: auto 1fr;
+  border-left: 1px solid rgba(255, 255, 255, 0.14);
+  background:
+    linear-gradient(160deg, rgba(15, 23, 42, 0.94), rgba(30, 41, 59, 0.88)),
+    rgba(15, 23, 42, 0.9);
+  box-shadow: -24px 0 80px rgba(0, 0, 0, 0.38);
+  color: #f8fafc;
+  backdrop-filter: blur(20px) saturate(1.2);
+}
+
+.settings-drawer-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 18px;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.12);
+  padding: 22px 22px 18px;
+}
+
+.settings-drawer-kicker {
+  display: block;
+  color: rgba(226, 232, 240, 0.62);
+  font-size: 12px;
+  font-weight: 800;
+  text-transform: uppercase;
+}
+
+.settings-drawer-header h1 {
+  margin: 2px 0 0;
+  color: #ffffff;
+  font-size: 24px;
+  font-weight: 800;
+}
+
+.drawer-close {
+  display: grid;
+  width: 40px;
+  height: 40px;
+  place-items: center;
+  border: 1px solid rgba(255, 255, 255, 0.14);
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.08);
+  color: #e2e8f0;
+  cursor: pointer;
+  font-size: 24px;
+  line-height: 1;
+}
+
+.drawer-close:hover {
+  background: rgba(125, 211, 252, 0.2);
+  color: #ffffff;
+}
+
+.settings-drawer-body {
+  display: grid;
+  align-content: start;
+  gap: 18px;
+  overflow: auto;
+  padding: 18px 22px 28px;
+}
+
+.settings-group {
+  display: grid;
+  gap: 12px;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  border-radius: 22px;
+  background: rgba(255, 255, 255, 0.07);
+  padding: 16px;
+}
+
+.settings-drawer-section {
+  display: grid;
+  gap: 14px;
+}
+
+.settings-drawer-section > h2 {
+  display: none;
+}
+
+.widget-settings-group h2:not(:first-child) {
+  margin-top: 12px;
+  border-top: 1px solid #e5e7eb;
+  padding-top: 16px;
+}
+
+.settings-group:last-child {
+  padding-bottom: 16px;
+}
+
+.settings-group h2 {
+  margin: 0;
+  color: #ffffff;
+  font-size: 16px;
+  font-weight: 900;
+}
+
+.settings-group label {
+  display: grid;
+  gap: 6px;
+}
+
+.settings-group label:has(input[type="checkbox"]) {
+  display: flex;
+  min-height: 32px;
+  align-items: center;
+  justify-content: space-between;
+  gap: 16px;
+}
+
+.settings-group label span {
+  color: rgba(226, 232, 240, 0.78);
+  font-size: 12px;
+  font-weight: 800;
+}
+
+.settings-group select,
+.settings-group input[type="range"] {
+  width: 100%;
+}
+
+.settings-group select {
+  height: 36px;
+  border: 1px solid #d1d5db;
+  border-radius: 6px;
+  background: #ffffff;
+  color: #111827;
+  padding: 0 9px;
+}
+
+.settings-group input[type="range"] {
+  accent-color: #7dd3fc;
+}
+
+.upload-button,
+.launcher-action {
+  display: flex;
+  min-height: 40px;
+  align-items: center;
+  justify-content: center;
+  border: 1px solid rgba(255, 255, 255, 0.14);
+  border-radius: 999px;
+  background: rgba(255, 255, 255, 0.08);
+  color: #f8fafc;
+  cursor: pointer;
+  font: inherit;
+  font-weight: 900;
+  padding: 0 14px;
+  text-align: center;
+}
+
+.upload-button:hover,
+.launcher-action:hover {
+  background: rgba(125, 211, 252, 0.2);
+  border-color: rgba(125, 211, 252, 0.42);
+}
+
+.settings-action-row {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 10px;
+}
+
+.upload-button input {
+  display: none;
+}
+
+.wallpaper-preview {
+  position: relative;
+  height: 126px;
+  overflow: hidden;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  border-radius: 18px;
+  background: rgba(15, 23, 42, 0.6);
+}
+
+.wallpaper-preview-image {
+  position: absolute;
+  inset: 0;
+  overflow: hidden;
+  background: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='800' height='420' viewBox='0 0 800 420'%3E%3Crect width='800' height='420' fill='%23d9ddd9'/%3E%3Cpath d='M0 316c90-46 160-51 241-16 80 34 155 26 234-5 127-50 215-36 325 25v100H0z' fill='%23222f2e'/%3E%3Cpath d='M0 256c77-41 151-60 228-40 85 22 131 60 231 30 117-35 224-69 341 4v170H0z' fill='%2332463e' opacity='.72'/%3E%3C/svg%3E");
+  background-size: cover;
+  background-position: center;
+}
+
+.wallpaper-preview-image::after {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background: rgba(0, 0, 0, var(--wallpaper-dim));
+}
+
+.wallpaper-preview-image img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  filter: blur(var(--wallpaper-blur));
+  transform: scale(1.04);
+}
+
+.launcher-message {
+  margin: 0;
+  color: rgba(226, 232, 240, 0.72);
+  font-size: 12px;
+  line-height: 1.35;
+  padding: 0 10px 4px;
+}
