+++ src/ui/shortcut-editor/index.ts
@@ -1,0 +1,1 @@
+export { ShortcutForm } from "./ShortcutForm";
