+++ src/ui/shortcut-editor/shortcut-form.css
@@ -1,0 +1,113 @@
+.quick-link-form {
+  display: grid;
+  gap: 16px;
+  padding: 18px 24px 24px;
+}
+
+.quick-link-form label {
+  display: grid;
+  gap: 7px;
+}
+
+.quick-link-form label span {
+  color: #4b5563;
+  font-size: 13px;
+  font-weight: 650;
+}
+
+.quick-link-form input {
+  width: 100%;
+  min-width: 0;
+  height: 44px;
+  border: 1px solid #d1d5db;
+  border-radius: 6px;
+  background: #ffffff;
+  color: #111827;
+  padding: 0 12px;
+}
+
+.quick-link-form input:focus {
+  border-color: #2563eb;
+  outline: 3px solid rgba(37, 99, 235, 0.18);
+}
+
+.form-row {
+  display: grid;
+  grid-template-columns: 1fr 118px;
+  gap: 14px;
+}
+
+.quick-link-form .color-input {
+  padding: 4px;
+}
+
+.form-preview {
+  display: grid;
+  justify-items: center;
+  padding: 4px 0;
+}
+
+.recommended-icons {
+  display: grid;
+  gap: 8px;
+}
+
+.recommended-icons > span {
+  color: #4b5563;
+  font-size: 13px;
+  font-weight: 650;
+}
+
+.recommended-icon-row {
+  display: grid;
+  grid-template-columns: repeat(6, 42px);
+  gap: 8px;
+}
+
+.recommended-icon-row button {
+  display: grid;
+  width: 42px;
+  height: 42px;
+  place-items: center;
+  border: 1px solid #d1d5db;
+  border-radius: 8px;
+  background: #f8fafc;
+  color: #111827;
+  cursor: pointer;
+}
+
+.recommended-icon-row button:hover,
+.recommended-icon-row button.selected {
+  border-color: #2563eb;
+  background: #eff6ff;
+  color: #1d4ed8;
+}
+
+.recommended-icon-row svg {
+  width: 20px;
+  height: 20px;
+  fill: currentColor;
+}
+
+.icon-image-actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  justify-content: center;
+}
+
+.file-button {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.file-button input {
+  display: none;
+}
+
+@media (max-width: 760px) {
+  .form-row {
+    grid-template-columns: 1fr;
+  }
+}
