+++ src/ui/widgets/WidgetContextMenu.css
@@ -1,0 +1,100 @@
+.widget-context-backdrop {
+  position: fixed;
+  z-index: 80;
+  inset: 0;
+}
+
+.widget-context-menu {
+  position: fixed;
+  display: grid;
+  width: min(320px, calc(100vw - 24px));
+  max-height: min(680px, calc(100vh - 24px));
+  gap: 12px;
+  overflow: auto;
+  border: 1px solid rgba(255, 255, 255, 0.18);
+  border-radius: 22px;
+  background:
+    linear-gradient(145deg, rgba(15, 23, 42, 0.92), rgba(30, 41, 59, 0.82)),
+    rgba(15, 23, 42, 0.84);
+  box-shadow:
+    0 28px 80px rgba(0, 0, 0, 0.42),
+    inset 0 1px 0 rgba(255, 255, 255, 0.14);
+  color: #f8fafc;
+  padding: 14px;
+  left: min(var(--context-menu-x), calc(100vw - min(320px, calc(100vw - 24px)) - 12px));
+  top: min(var(--context-menu-y), calc(100vh - min(680px, calc(100vh - 24px)) - 12px));
+  backdrop-filter: blur(18px) saturate(1.25);
+}
+
+.widget-context-header {
+  display: grid;
+  gap: 2px;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.12);
+  padding: 4px 2px 12px;
+}
+
+.widget-context-header span {
+  color: rgba(226, 232, 240, 0.66);
+  font-size: 11px;
+  font-weight: 900;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+}
+
+.widget-context-header strong {
+  color: #ffffff;
+  font-size: 18px;
+  font-weight: 900;
+}
+
+.widget-context-menu label {
+  display: grid;
+  gap: 7px;
+  border-radius: 16px;
+  background: rgba(255, 255, 255, 0.06);
+  padding: 10px 12px;
+}
+
+.widget-context-menu label span,
+.widget-context-menu output {
+  color: rgba(241, 245, 249, 0.86);
+  font-size: 12px;
+  font-weight: 800;
+}
+
+.context-toggle-row {
+  display: flex !important;
+  min-height: 42px;
+  align-items: center;
+  justify-content: space-between;
+  gap: 14px;
+}
+
+.context-range-row {
+  grid-template-columns: 1fr auto;
+}
+
+.context-range-row input {
+  grid-column: 1 / -1;
+}
+
+.widget-context-menu input[type="range"] {
+  width: 100%;
+  accent-color: #7dd3fc;
+}
+
+.widget-context-menu input[type="checkbox"] {
+  width: 18px;
+  height: 18px;
+  accent-color: #7dd3fc;
+}
+
+.widget-context-menu select {
+  width: 100%;
+  height: 36px;
+  border: 1px solid rgba(255, 255, 255, 0.16);
+  border-radius: 12px;
+  background: rgba(15, 23, 42, 0.92);
+  color: #f8fafc;
+  padding: 0 10px;
+}
