+++ src/ui/widgets/search/index.ts
@@ -1,0 +1,2 @@
+export { SearchWidget } from "./SearchWidget";
+export { SearchWidgetContextMenu } from "./SearchWidgetContextMenu";
