+++ src/ui/widgets/shortcut-grid/index.ts
@@ -1,0 +1,3 @@
+export { FolderPanel } from "./FolderPanel";
+export { ShortcutGridWidget } from "./ShortcutGridWidget";
+export { ShortcutGridWidgetContextMenu } from "./ShortcutGridWidgetContextMenu";
