+++ src/ui/widgets/shortcut-grid/shortcuts/index.ts
@@ -1,0 +1,1 @@
+export { ShortcutIcon } from "./ShortcutIcon";
