+++ src/ui/widgets/shortcut-grid/shortcuts/shortcut-tiles.css
@@ -1,0 +1,234 @@
+.quick-link {
+  display: grid;
+  grid-template-rows: auto auto;
+  width: 100%;
+  min-width: 0;
+  min-height: 0;
+  align-content: start;
+  justify-items: center;
+  gap: var(--quick-link-tile-gap);
+  color: #ffffff;
+  text-align: center;
+  text-decoration: none;
+  position: relative;
+  transition:
+    opacity 140ms ease,
+    transform 140ms ease;
+}
+
+.quick-link[draggable="true"] {
+  cursor: grab;
+}
+
+.quick-link.dragging-origin {
+  opacity: 0.5 !important;
+}
+
+.quick-link.drag-over .quick-link-icon {
+  outline: 3px solid rgba(255, 255, 255, 0.88);
+  outline-offset: 7px;
+}
+
+.quick-link.drop-center .quick-link-icon {
+  transform: scale(0.92);
+  animation: combine-pulse 700ms ease-in-out infinite;
+  box-shadow: 0 0 20px rgba(99, 102, 241, 0.6);
+}
+
+@keyframes combine-pulse {
+  0%, 100% {
+    transform: scale(0.92);
+    box-shadow: 0 0 20px rgba(99, 102, 241, 0.6);
+  }
+
+  50% {
+    transform: scale(1);
+    box-shadow: 0 0 30px rgba(99, 102, 241, 0.8);
+  }
+}
+
+.quick-link.drop-leading .quick-link-icon,
+.quick-link.drop-trailing .quick-link-icon {
+  box-shadow: 0 18px 34px rgba(0, 0, 0, 0.24);
+}
+
+.quick-link.combine-preview .quick-link-icon {
+  transform: scale(0.85);
+  box-shadow: 0 0 15px rgba(99, 102, 241, 0.5);
+}
+
+.drag-overlay-tile {
+  display: grid;
+  justify-items: center;
+  gap: var(--quick-link-tile-gap);
+  color: #ffffff;
+  text-align: center;
+  text-decoration: none;
+  pointer-events: none;
+  position: relative;
+  filter: drop-shadow(0 12px 24px rgba(0, 0, 0, 0.4));
+  z-index: 50;
+  transition: transform 140ms ease;
+}
+
+.drag-overlay-tile.merge-active .quick-link-icon {
+  transform: scale(0.85);
+}
+
+.drag-overlay-merge {
+  position: absolute;
+  top: -8px;
+  right: -8px;
+  display: grid;
+  width: 30px;
+  height: 30px;
+  place-items: center;
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.94);
+  color: #1f2937;
+  font-size: 22px;
+  font-weight: 900;
+  line-height: 1;
+}
+
+.quick-link-icon {
+  display: grid;
+  width: var(--icon-size);
+  height: var(--icon-size);
+  place-items: center;
+  border-radius: 50%;
+  color: #ffffff;
+  font-size: calc(var(--icon-size) * 0.42);
+  font-weight: 800;
+  overflow: hidden;
+  transition:
+    outline-color 140ms ease,
+    outline-offset 140ms ease,
+    transform 140ms ease;
+}
+
+.quick-link-icon img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.quick-link-icon svg {
+  width: 52%;
+  height: 52%;
+  fill: #ffffff;
+}
+
+.quick-link-icon.image-icon {
+  background: #111827;
+}
+
+.quick-link:hover .quick-link-icon {
+  transform: translateY(-2px);
+}
+
+.quick-link-title {
+  max-width: 100%;
+  overflow: hidden;
+  color: rgba(255, 255, 255, 0.9);
+  font-size: var(--quick-link-label-font-size);
+  line-height: 1.2;
+  text-overflow: ellipsis;
+  text-shadow: 0 1px 3px rgba(0, 0, 0, 0.42);
+  white-space: nowrap;
+}
+
+.quick-link-edit {
+  position: absolute;
+  top: -8px;
+  right: calc(50% - var(--icon-size) / 2 - 8px);
+  border: 0;
+  border-radius: 999px;
+  background: rgba(255, 255, 255, 0.96);
+  color: #1f2937;
+  cursor: pointer;
+  font-size: 12px;
+  line-height: 1;
+  opacity: 0;
+  padding: 6px 9px;
+  pointer-events: none;
+  transform: translateY(4px);
+  transition:
+    opacity 150ms ease,
+    transform 150ms ease;
+}
+
+.quick-link:hover .quick-link-edit,
+.quick-link:focus-within .quick-link-edit {
+  opacity: 1;
+  pointer-events: auto;
+  transform: translateY(0);
+}
+
+.add-link,
+.folder-link {
+  border: 0;
+  background: transparent;
+  cursor: pointer;
+  font: inherit;
+}
+
+.add-link-icon {
+  border: 2px dashed rgba(255, 255, 255, 0.72);
+  background: rgba(255, 255, 255, 0.16);
+  color: rgba(255, 255, 255, 0.92);
+  font-size: calc(var(--icon-size) * 0.5);
+  font-weight: 300;
+}
+
+.folder-icon {
+  position: relative;
+}
+
+.folder-preview-icon {
+  background: rgba(255, 255, 255, 0.16);
+  border: 1px solid rgba(255, 255, 255, 0.22);
+  backdrop-filter: none;
+  box-shadow: none;
+  overflow: visible;
+}
+
+.folder-preview-grid {
+  display: grid;
+  width: 58%;
+  grid-template-columns: repeat(3, 1fr);
+  gap: max(2px, calc(var(--icon-size) * 0.045));
+}
+
+.folder-preview-child {
+  display: grid;
+  aspect-ratio: 1;
+  min-width: 0;
+  place-items: center;
+  border-radius: 50%;
+  color: #ffffff;
+  font-size: max(6px, calc(var(--icon-size) * 0.075));
+  font-weight: 900;
+  line-height: 1;
+  overflow: hidden;
+}
+
+.folder-preview-child img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.folder-preview-child svg {
+  width: 58%;
+  height: 58%;
+  fill: #ffffff;
+}
+
+@media (max-width: 760px) {
+  .quick-link-edit {
+    opacity: 1;
+    pointer-events: auto;
+    transform: none;
+  }
+}
