+++ src/ui/widgets/shortcut-grid/useNativeDragOverlay.ts
@@ -1,0 +1,36 @@
+import { useEffect } from "react";
+import type { ResolvedTopLevelTile } from "../../../domain/tabOperations";
+
+export type DragOverlayState = {
+  tile: ResolvedTopLevelTile;
+  x: number;
+  y: number;
+} | null;
+
+export function useNativeDragOverlayPointer(
+  dragOverlay: DragOverlayState,
+  setDragOverlay: (overlay: DragOverlayState | ((current: DragOverlayState) => DragOverlayState)) => void
+) {
+  // Native drag emits pointer position through dragover; this hook isolates that browser-specific Adapter.
+  useEffect(() => {
+    if (!dragOverlay) return;
+
+    const handleMove = (event: MouseEvent) => {
+      setDragOverlay((current) => (current ? { ...current, x: event.clientX, y: event.clientY } : null));
+    };
+
+    const handleEnd = () => {
+      setDragOverlay(null);
+    };
+
+    window.addEventListener("dragover", handleMove);
+    window.addEventListener("dragend", handleEnd);
+    window.addEventListener("drop", handleEnd);
+
+    return () => {
+      window.removeEventListener("dragover", handleMove);
+      window.removeEventListener("dragend", handleEnd);
+      window.removeEventListener("drop", handleEnd);
+    };
+  }, [dragOverlay, setDragOverlay]);
+}
