+++ src/ui/widgets/widgets.css
@@ -1,0 +1,5 @@
+@import "./widget-frame.css";
+@import "./WidgetContextMenu.css";
+@import "./search/search-widget.css";
+@import "./shortcut-grid/shortcut-grid-widget.css";
+@import "./shortcut-grid/shortcuts/shortcut-tiles.css";
